use a base project page model
