+++ src/Frontend/Dfe.Complete/Pages/Projects/List/AllProjectsModel.cs
@@ -1,15 +1,10 @@
 ﻿using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Constants;
-using Dfe.Complete.Domain.Enums;
-using Dfe.Complete.Domain.ValueObjects;
 using Dfe.Complete.Models;
-using Dfe.Complete.Pages.Pagination;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace Dfe.Complete.Pages.Projects.List;
 
-public abstract class AllProjectsModel(string currentNavigation) : PageModel
+public abstract class AllProjectsModel(string currentNavigation) : BaseProjectsPageModel(currentNavigation)
 {
     protected TabNavigationModel AllProjectsTabNavigationModel = new(TabNavigationModel.AllProjectsTabName);
 
@@ -23,30 +18,6 @@ public abstract class AllProjectsModel(string currentNavigation) : PageModel
     public const string CompletedNavigation = "Completed";
     public const string StatisticsNavigation = "Statistics";
 
-    public string CurrentNavigationItem { get; init; } = currentNavigation;
-
-    [BindProperty(SupportsGet = true)] public int PageNumber { get; set; } = 1;
-
-    public PaginationModel? Pagination { get; set; }
-
-    internal int PageSize = 20;
-
-    public static string GetProjectSummaryUrl(ListAllProjectsResultModel project)
-    {
-        return string.Format(
-            project.ProjectType == ProjectType.Conversion
-                ? RouteConstants.ConversionProjectTaskList
-                : RouteConstants.TransferProjectTaskList, project.ProjectId.Value);
-    }
-
-    public static string GetProjectSummaryUrl(ProjectType? type, ProjectId projectId)
-    {
-        return string.Format(
-            type == ProjectType.Conversion
-                ? RouteConstants.ConversionProjectTaskList
-                : RouteConstants.TransferProjectTaskList, projectId.Value);
-    }
-    
     public static string GetTrustProjectsUrl(ListTrustsWithProjectsResultModel trustModel)
     {
         return trustModel.identifier.Contains("TR") ?
